feat: is connected

+++ main.cpp
@@ -4212,7 +4212,8 @@ GValue phpgtk_get_gvalue(Php::Value phpgtk_value, GType type_column)
         case G_TYPE_LONG:
         {
             // Cast
-            long b = (long long)phpgtk_value;
+
+            long long b = (long long)phpgtk_value;
 
             g_value_init(&gtk_value, G_TYPE_LONG);
             g_value_set_long(&gtk_value, b);
@@ -4221,10 +4222,11 @@ GValue phpgtk_get_gvalue(Php::Value phpgtk_value, GType type_column)
         case G_TYPE_ULONG:
         {
             // Cast
-            //int b = (int64_t)phpgtk_value;
+            // unsigned long long b = (unsigned long long)phpgtk_value;
 
-            //g_value_init(&gtk_value, G_TYPE_ULONG);
-            //g_value_set_ulong(&gtk_value, b);
+            // g_value_init(&gtk_value, G_TYPE_ULONG);
+            // g_value_set_ulong(&gtk_value, b);
+           // throw Php::Exception("G_TYPE_ULONG not implemented");
             break;
         }
         case G_TYPE_UINT:
